Move file and application tests under views directory to match formatting of nodes and users
